requirements.txt
appnope==0.1.0
asgiref==3.2.5
attrs==19.3.0
backcall==0.1.0
bibtexparser==1.1.0
bleach==3.1.3
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.4
django-audit-log==0.7.0
django-autoslug==1.9.8
django-el-pagination==3.3.0
django-filter==2.2.0
django-icons==1.1.2
django-markdownx==3.0.1
django-simple-history==2.8.0
entrypoints==0.3
feedparser==5.2.1
future==0.18.2
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Markdown==3.2.1
MarkupSafe==1.1.1
martor==1.4.9
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
parso==0.6.2
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
pyzmq==19.0.0
Pyzotero==1.4.15
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
